backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
aiomysql==0.2.0
alembic==1.12.1
pymysql==1.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
PyJWT==2.8.0

# Google OAuth
httpx==0.25.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Stripe
stripe==7.5.0

# WebSocket
websockets==12.0

# CORS
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code quality
ruff==0.1.7
mypy==1.7.1

# Development
ipython==8.18.1

# Monitoring and logging
structlog==23.2.0

# Redis for caching (optional)
redis==5.0.1